// src/components/schedular/types.d.ts
interface CalendarEvent {
    event_id: number | string;
    title: string;
    start: Date;
    end: Date;
    disabled?: boolean;
    color?: string;
    editable?: boolean;
    deletable?: boolean;
    draggable?: boolean;
    allDay?: boolean;
}
// ...
export type ProcessedEvent = CalendarEvent & Record<string, any>;
export type EventActions = "create" | "edit";
// ...
export interface SchedulerHelpers {
    state: Record<string, StateItem>;
    close(): void;
    loading(status: boolean): void;
    edited?: ProcessedEvent;
    onConfirm(event: ProcessedEvent | ProcessedEvent[], action: EventActions): void;
    [resourceKey: string]: unknown;
}